refactor(alerts): split alert manager state
Move alert state persistence and alert settings models into a focused module while preserving alert_manager imports for existing callers and tests.

Tested: pytest tests/test_alert_manager.py tests/test_alert_notification_system.py tests/test_alert_lifecycle.py tests/test_alert_lifecycle_pipeline.py tests/test_alert_lifecycle_presentation.py; ruff check src tests scripts; pyright.

Confidence: high
