+++ src/accessiweather/alert_manager.py
@@ -8,243 +8,17 @@ alert identification and user controls.
 
 import logging
 import time
-from collections import deque
 from datetime import UTC, datetime, timedelta
 from pathlib import Path
 
+from .alert_manager_state import AlertSettings, AlertState
 from .constants import (
-    ALERT_HISTORY_MAX_LENGTH,
-    DEFAULT_ESCALATION_COOLDOWN_MINUTES,
-    DEFAULT_FRESHNESS_WINDOW_MINUTES,
-    DEFAULT_GLOBAL_COOLDOWN_MINUTES,
-    DEFAULT_MAX_NOTIFICATIONS_PER_HOUR,
-    DEFAULT_MIN_SEVERITY_PRIORITY,
-    DEFAULT_NOTIFICATIONS_ENABLED,
-    DEFAULT_PER_ALERT_COOLDOWN_MINUTES,
-    DEFAULT_SOUND_ENABLED,
     SECONDS_PER_HOUR,
-    SEVERITY_PRIORITY_MAP,
 )
 from .models import WeatherAlert, WeatherAlerts
 from .runtime_state import RuntimeStateManager
 
 logger = logging.getLogger(__name__)
-
-
-class AlertState:
-    """
-    Represents the state of a tracked alert with history tracking.
-
-    Maintains a bounded history of content changes (hash, priority, timestamp)
-    to enable accurate change detection and escalation tracking.
-    """
-
-    def __init__(
-        self,
-        alert_id: str,
-        content_hash: str,
-        first_seen: datetime,
-        last_notified: datetime | None = None,
-        notification_count: int = 0,
-        severity_priority: int = 1,
-        alert_sent_time: datetime | None = None,
-    ):
-        """
-        Initialize alert state.
-
-        Args:
-        ----
-            alert_id: Unique identifier for the alert
-            content_hash: Hash of alert content for change detection
-            first_seen: Timestamp when alert was first seen
-            last_notified: Timestamp of last notification sent
-            notification_count: Number of times notifications were sent
-            severity_priority: Numeric severity priority (1-5)
-            alert_sent_time: Timestamp when alert was issued by the provider (sent/effective)
-
-        """
-        self.alert_id = alert_id
-        self.first_seen = first_seen
-        self.last_notified = last_notified
-        self.notification_count = notification_count
-        self.alert_sent_time = alert_sent_time
-
-        # Bounded history: (content_hash, severity_priority, timestamp)
-        self.hash_history: deque[tuple[str, int, float]] = deque(maxlen=ALERT_HISTORY_MAX_LENGTH)
-        # Add initial state to history
-        self.add_hash(content_hash, severity_priority)
-
-    @property
-    def content_hash(self) -> str:
-        """Get current content hash (most recent in history)."""
-        if not self.hash_history:
-            return ""
-        return self.hash_history[-1][0]
-
-    def add_hash(self, content_hash: str, severity_priority: int, timestamp: float | None = None):
-        """
-        Add a new content hash to history.
-
-        Args:
-        ----
-            content_hash: Hash of alert content
-            severity_priority: Numeric severity priority (1-5)
-            timestamp: Unix timestamp (defaults to current time)
-
-        """
-        if timestamp is None:
-            timestamp = time.time()
-        self.hash_history.append((content_hash, severity_priority, timestamp))
-
-    def has_changed(self, new_hash: str) -> bool:
-        """
-        Check if content has changed from most recent state.
-
-        Args:
-        ----
-            new_hash: New content hash to compare
-
-        Returns:
-        -------
-            True if content has changed
-
-        """
-        if not self.hash_history:
-            return True
-        return self.hash_history[-1][0] != new_hash
-
-    def is_escalated(self, new_priority: int) -> bool:
-        """
-        Check if alert has escalated in severity.
-
-        Compares new priority against the highest priority seen in history.
-
-        Args:
-        ----
-            new_priority: New severity priority to check
-
-        Returns:
-        -------
-            True if new priority is higher than any in history
-
-        """
-        if not self.hash_history:
-            return False
-
-        # Get highest priority from history
-        max_historical_priority = max(entry[1] for entry in self.hash_history)
-        return new_priority > max_historical_priority
-
-    def get_previous_priority(self) -> int:
-        """
-        Get the previous severity priority.
-
-        Returns
-        -------
-            Previous priority, or 1 (unknown) if no history
-
-        """
-        if len(self.hash_history) < 2:
-            return 1  # Default to unknown priority
-        return self.hash_history[-2][1]
-
-    def to_dict(self) -> dict:
-        """Convert to dictionary for JSON serialization."""
-        return {
-            "alert_id": self.alert_id,
-            "first_seen": self.first_seen.isoformat(),
-            "last_notified": self.last_notified.isoformat() if self.last_notified else None,
-            "notification_count": self.notification_count,
-            "alert_sent_time": self.alert_sent_time.isoformat() if self.alert_sent_time else None,
-            # Store history as list of [hash, priority, timestamp]
-            "hash_history": [[h, p, t] for h, p, t in self.hash_history],
-        }
-
-    @classmethod
-    def from_dict(cls, data: dict) -> "AlertState":
-        """
-        Create from dictionary loaded from JSON.
-
-        Handles both old format (single content_hash) and new format (hash_history).
-        """
-        alert_id = data["alert_id"]
-        first_seen = datetime.fromisoformat(data["first_seen"])
-        last_notified = (
-            datetime.fromisoformat(data["last_notified"]) if data.get("last_notified") else None
-        )
-        notification_count = data.get("notification_count", 0)
-        alert_sent_time = (
-            datetime.fromisoformat(data["alert_sent_time"]) if data.get("alert_sent_time") else None
-        )
-
-        # Handle migration from old format
-        if "hash_history" in data:
-            # New format with history
-            state = cls(
-                alert_id=alert_id,
-                content_hash="",  # Will be replaced by history
-                first_seen=first_seen,
-                last_notified=last_notified,
-                notification_count=notification_count,
-                severity_priority=1,
-                alert_sent_time=alert_sent_time,
-            )
-            # Clear the initial entry and load history
-            state.hash_history.clear()
-            for entry in data["hash_history"]:
-                h, p, t = entry
-                state.hash_history.append((h, p, t))
-        else:
-            # Old format with single content_hash - migrate
-            content_hash = data.get("content_hash", "")
-            state = cls(
-                alert_id=alert_id,
-                content_hash=content_hash,
-                first_seen=first_seen,
-                last_notified=last_notified,
-                notification_count=notification_count,
-                severity_priority=1,  # Unknown priority for migrated data
-                alert_sent_time=alert_sent_time,
-            )
-
-        return state
-
-
-class AlertSettings:
-    """Alert notification settings and preferences."""
-
-    def __init__(self):
-        """Initialize the instance."""
-        # Severity threshold (minimum severity to notify)
-        self.min_severity_priority = DEFAULT_MIN_SEVERITY_PRIORITY
-
-        # Cooldown periods (in minutes)
-        self.global_cooldown = DEFAULT_GLOBAL_COOLDOWN_MINUTES
-        self.per_alert_cooldown = DEFAULT_PER_ALERT_COOLDOWN_MINUTES
-        self.escalation_cooldown = DEFAULT_ESCALATION_COOLDOWN_MINUTES
-        self.freshness_window_minutes = DEFAULT_FRESHNESS_WINDOW_MINUTES
-
-        # Category filters (event types to ignore)
-        self.ignored_categories: set[str] = set()
-
-        # General settings
-        self.notifications_enabled = DEFAULT_NOTIFICATIONS_ENABLED
-        self.sound_enabled = DEFAULT_SOUND_ENABLED
-        self.max_notifications_per_hour = DEFAULT_MAX_NOTIFICATIONS_PER_HOUR
-
-    def should_notify_severity(self, severity: str) -> bool:
-        """Check if we should notify for this severity level."""
-        if not self.notifications_enabled:
-            return False
-
-        priority = SEVERITY_PRIORITY_MAP.get(severity.lower(), SEVERITY_PRIORITY_MAP["unknown"])
-        return priority >= self.min_severity_priority
-
-    def should_notify_category(self, event: str) -> bool:
-        """Check if we should notify for this event category."""
-        if not event:
-            return True
-        return event.lower() not in self.ignored_categories
 
 
 class AlertManager:
